name cleanup and splitting main from modules
