+++ src/cpp/NT.cpp
@@ -3,13 +3,7 @@
 // github: https://github.com/JannesSP
 // website: https://jannessp.github.io
 
-#include "dynamont_NT.hpp"
-
-bool rna;
-std::unordered_map<std::string, double> transitions_NT = {
-    {"m1", -1.0},
-    {"e1", -1.0},
-    {"e2", -1.0}};
+#include "NT.hpp"
 
 /**
  * Computes forward matrices using logarithmic values for signal processing.
@@ -27,7 +21,7 @@ std::unordered_map<std::string, double> transitions_NT = {
  * @param N Length of nucleotide sequence + 1.
  * @param model Vector containing kmers as keys and (mean, stdev) tuples as values.
  */
-void logF(const double *sig, const int *kmerSeq, double *M, double *E, const std::size_t T, const std::size_t N, const std::tuple<double, double> *model)
+void logF_NT(const double *sig, const int *kmerSeq, double *M, double *E, const std::size_t T, const std::size_t N, const std::tuple<double, double> *model)
 {
     const double m1 = transitions_NT.at("m1");
     const double e2 = transitions_NT.at("e2");
@@ -40,7 +34,7 @@ void logF(const double *sig, const int *kmerSeq, double *M, double *E, const std
         for (std::size_t n = 1; n < N; ++n)
         {                                                                      // speed up, due to rules no need to look at upper triangle of matrices
             const double score = scoreKmer(sig[t - 1], kmerSeq[n - 1], model); // Cache scoreKmer for (t-1, n-1)
-            M[tN + n] = E[tN - N + (n - 1)] + score + m1;
+            M[tN + n] = E[tN - N + n - 1] + score + m1;
             E[tN + n] = logPlus(M[tN - N + n] + score, E[tN - N + n] + score + e2);
         }
     }
@@ -62,7 +56,7 @@ void logF(const double *sig, const int *kmerSeq, double *M, double *E, const std
  * @param N Length of nucleotide sequence + 1.
  * @param model Vector containing kmers as keys and (mean, stdev) tuples as values.
  */
-void logB(const double *sig, const int *kmerSeq, double *M, double *E, const std::size_t T, const std::size_t N, const std::tuple<double, double> *model)
+void logB_NT(const double *sig, const int *kmerSeq, double *M, double *E, const std::size_t T, const std::size_t N, const std::tuple<double, double> *model)
 {
     const double m1 = transitions_NT.at("m1");
     const double e2 = transitions_NT.at("e2");
@@ -87,27 +81,6 @@ void logB(const double *sig, const int *kmerSeq, double *M, double *E, const std
             }
             E[tN + n] = ext;
         }
-    }
-}
-
-/**
- * Calculate the logarithmic probability matrix
- *
- * @param LP Matrix to store the logarithmic probabilities.
- * @param FOR Matrix containing forward-values for segment borders.
- * @param BACK Matrix containing backward-values for extending segment.
- * @param Z Alignment score.
- * @param S Size of matrix.
- *
- * This function calculates the logarithmic probability matrix for a given forward and backward
- * matrix by adding the values of the forward and backward matrix and subtracting the alignment score.
- */
-void logP(double *LP, const double *FOR, const double *BACK, const double Z, const std::size_t S)
-{
-#pragma omp parallel for
-    for (std::size_t s = 0; s < S; ++s)
-    {
-        LP[s] = FOR[s] + BACK[s] - Z;
     }
 }
 
@@ -203,69 +176,6 @@ void traceback(std::size_t t, std::size_t n, const double *M, const double *E, c
         }
     }
 }
-
-/**
- * Backtracing function for M state
- *
- * @param t current time step
- * @param n current position in sequence
- * @param M matrix containing match scores
- * @param E matrix containing extend scores
- * @param LPM matrix containing logarithmic probabilities for M state
- * @param LPE matrix containing logarithmic probabilities for E state
- * @param segString list to store segment information
- * @param N size of sequence
- * @param segProb vector to store segment probabilities
- * @param kmerSize size of k-mer
- */
-// void funcM(const std::size_t t, const std::size_t n, const double *M, const double *E, const double *LPM, const double *LPE, std::list<std::string> &segString, const std::size_t N, std::vector<double> &segProb, const int kmerSize)
-// {
-//     const double score = M[t * N + n];
-//     const double logScore = LPM[t * N + n];
-//     segProb.push_back(exp(logScore));
-//     // if (t<=1 && n<=1) [[unlikely]] { // Start value
-//     //     segString.push_front("M0,0," + formattedMedian(segProb) + ";"); // n-1 because N is 1 larger than the sequences
-//     //     return;
-//     // }
-//     if (t > 0 && n > 0 && score == E[(t - 1) * N + (n - 1)] + logScore) [[likely]]
-//     {
-//         segString.push_front("M" + std::to_string(n - 1 + kmerSize / 2) + "," + std::to_string(t - 1) + "," + formattedMedian(segProb) + ";");
-//         segProb.clear();
-//         return funcE(t - 1, n - 1, M, E, LPM, LPE, segString, N, segProb, kmerSize);
-//     }
-// }
-
-/**
- * Backtracing function for E state
- *
- * @param t current time step
- * @param n current position in sequence
- * @param M matrix containing match scores
- * @param E matrix containing extend scores
- * @param LPM matrix containing logarithmic probabilities for M state
- * @param LPE matrix containing logarithmic probabilities for E state
- * @param segString list to store segment information
- * @param N size of sequence
- * @param segProb vector to store segment probabilities
- * @param kmerSize size of k-mer
- */
-// void funcE(const std::size_t t, const std::size_t n, const double *M, const double *E, const double *LPM, const double *LPE, std::list<std::string> &segString, const std::size_t N, std::vector<double> &segProb, const int kmerSize)
-// {
-//     const double score = E[t * N + n];
-//     const double logScore = LPE[t * N + n];
-//     segProb.push_back(exp(logScore));
-//     if (t > 0 && n > 0) [[likely]]
-//     {
-//         if (score == M[(t - 1) * N + n] + logScore)
-//         {
-//             return funcM(t - 1, n, M, E, LPM, LPE, segString, N, segProb, kmerSize);
-//         }
-//         if (score == E[(t - 1) * N + n] + logScore)
-//         {
-//             return funcE(t - 1, n, M, E, LPM, LPE, segString, N, segProb, kmerSize);
-//         }
-//     }
-// }
 
 /**
  * Train transition parameter with baum welch algorithm
@@ -466,267 +376,3 @@ void trainParams(const double *sig, const int *kmerSeq, const double *forM, cons
     delete[] newMeans;
     delete[] newStdevs;
 }
-
-/**
- * Main entry point for the dynamont command line tool.
- *
- * @param argc Number of arguments passed to the program.
- * @param argv Array of strings containing the arguments passed to the program.
- *
- * @return 0 if the program succeeded, 1 if there was an error parsing the arguments, 2 if the input was invalid, 3 if the pore was not supported, 11 if the Z values did not match between forward and backward matrices.
- *
- * @details
- * This function parses the command line arguments, loads the model, reads the input, calculates the segmentation probabilities and prints the result.
- */
-#ifndef UNIT_TESTING
-int main(int argc, char *argv[])
-{
-    // speedup for I/O
-    std::ios_base::sync_with_stdio(0);
-    std::cin.tie(0);
-    std::cout.tie(0);
-
-    bool train, calcZ, prob;
-    std::string pore, modelpath;
-
-    std::cerr << std::fixed << std::showpoint << std::setprecision(5);
-    std::cout << std::fixed << std::showpoint << std::setprecision(5);
-
-    // Argparser
-    argparse::ArgumentParser program("dynamont basic", "0.1");
-    // parameters for DP
-    program.add_argument("-m", "--model").help("Path to kmer model table").required().store_into(modelpath);
-    program.add_argument("-r", "--pore").help("Pore used to sequence the data").required().choices("rna_r9", "dna_r9", "rna_rp4", "dna_r10_260bps", "dna_r10_400bps").store_into(pore);
-
-    program.add_argument("-m1", "--matchscore1").help("Segment transition probability, should be close to (expected number of nucleotdes)/(signal length). Leave at -1 if unset.").default_value(-1.0).scan<'g', double>().store_into(transitions_NT["m1"]);
-    program.add_argument("-e1", "--extendscore1").help("First extend probability.").default_value(-1.0).scan<'g', double>().store_into(transitions_NT["e1"]);
-    program.add_argument("-e2", "--extendscore2").help("Further extend probability.").default_value(-1.0).scan<'g', double>().store_into(transitions_NT["e2"]);
-    program.add_argument("--train").help("Switch algorithm to transition and emission parameter training mode").default_value(false).implicit_value(true).store_into(train);
-    program.add_argument("-z", "--calcZ").help("Switch algorithm to only calculate Z").default_value(false).implicit_value(true).store_into(calcZ);
-    program.add_argument("-p", "--probabilty").help("Print out the segment border probability").default_value(false).implicit_value(true).store_into(prob);
-    program.add_argument("-t").help("Number of threads to use").default_value(1).scan<'i', int>();
-
-    program.parse_args(argc, argv);
-
-    omp_set_dynamic(1);
-    omp_set_num_threads(program.get<int>("-t"));
-
-    int kmerSize = 0;
-    // load default and set parameters
-    if (pore == "rna_r9")
-    {
-        kmerSize = 5;
-        rna = true;
-        // taken from the trained NT version of dynamont
-        updateTransitions(NT_rna_r9_transitions, transitions_NT);
-    }
-    else if (pore == "dna_r9")
-    {
-        kmerSize = 5;
-        rna = false;
-        // taken from the trained NT version of dynamont
-        updateTransitions(NT_dna_r9_transitions, transitions_NT);
-    }
-    else if (pore == "rna_rp4")
-    {
-        kmerSize = 9;
-        rna = true;
-        // taken from the trained NT version of dynamont
-        updateTransitions(NT_rna_rp4_transitions, transitions_NT);
-    }
-    else if (pore == "dna_r10_260bps")
-    {
-        kmerSize = 9;
-        rna = false;
-        updateTransitions(NT_dna_r10_260bps_transitions, transitions_NT);
-    }
-    else if (pore == "dna_r10_400bps")
-    {
-        kmerSize = 9;
-        rna = false;
-        updateTransitions(NT_dna_r10_400bps_transitions, transitions_NT);
-    }
-
-    checkModelpath(modelpath);
-    auto result = readKmerModel(modelpath, kmerSize, rna);
-    std::tuple<double, double> *model = std::get<0>(result);
-    const int alphabetSize = std::get<1>(result);
-    const int numKmers = std::get<2>(result);
-
-    // example
-    // 107,107,107.2,108.0,108.9,111.2,105.7,104.3,107.1,105.7,105,105
-    // CAAAAA
-    // read input, signal and read whitespace separated in single line
-    std::string signal, read;
-    getline(std::cin, signal);
-    getline(std::cin, read);
-
-    if (signal.empty())
-    {
-        std::cout << "Signal missing!" << std::endl;
-        exit(4);
-    }
-    else if (read.empty())
-    {
-        std::cout << "Read missing!" << std::endl;
-        exit(5);
-    }
-
-    // process signal: convert std::string to double std::array
-    const std::size_t T = count(signal.begin(), signal.end(), ',') + 2; // len(sig) + 1
-    checkInput(T, read.size(), kmerSize);
-    const std::size_t N = read.size() - kmerSize + 1 + 1; // N is number of kmers in sequence + 1
-    const std::size_t TN = T * N;
-
-    double *sig = new double[T - 1];
-    std::string value;
-    std::stringstream ss(signal);
-    int i = 0;
-    while (getline(ss, value, ','))
-    {
-        sig[i++] = stod(value);
-    }
-
-    // process read N: convert std::string to int std::array
-    int *kmerSeq = new int[N - 1];
-#pragma omp parallel for
-    for (std::size_t n = 0; n < N - 1; ++n)
-    {
-        kmerSeq[n] = kmer2int(read.substr(n, kmerSize), alphabetSize);
-    }
-
-    // deallocate memory
-    ss.clear();
-    signal.erase();
-    read.erase();
-    value.erase();
-
-    // std::cerr<<"T: "<<T<<", "<<"N: "<<N<<", "<<"inputsize: "<<TN<<std::endl;
-    // calculate segmentation probabilities, fill forward matrices
-    double *forM = new double[TN];
-    double *forE = new double[TN];
-    // calculate segmentation probabilities, fill backward matrices
-    double *backM = new double[TN];
-    double *backE = new double[TN];
-#pragma omp parallel for
-    for (std::size_t i = 0; i < TN; ++i)
-    {
-        forM[i] = -INFINITY;
-        forE[i] = -INFINITY;
-        backM[i] = -INFINITY;
-        backE[i] = -INFINITY;
-    }
-    logF(sig, kmerSeq, forM, forE, T, N, model);
-    logB(sig, kmerSeq, backM, backE, T, N, model);
-    const double Zf = forE[TN - 1];
-    const double Zb = backE[0];
-
-    // Numeric error is scaled by input size, Z in forward and backward should match by some numeric error EPSILON
-    if (abs(Zf - Zb) / TN > EPSILON || std::isinf(Zf) || std::isinf(Zb))
-    {
-        std::cerr << "Z values between matrices do not match! Zf: " << Zf << ", Zb: " << Zb << ", " << abs(Zf - Zb) / TN << " > " << EPSILON << std::endl;
-        delete[] sig;
-        delete[] kmerSeq;
-        delete[] forM;
-        delete[] forE;
-        delete[] backM;
-        delete[] backE;
-        delete[] model;
-        exit(3);
-    }
-
-    // std::cerr<<"Zf: "<<Zf<<", Zb: "<<Zb<<", "<<abs(Zf-Zb)/TN<<" <! "<<EPSILON<<std::endl;
-
-    if (calcZ)
-    {
-        std::cout << Zb << std::endl;
-        delete[] sig;
-        delete[] kmerSeq;
-        delete[] forM;
-        delete[] forE;
-        delete[] backM;
-        delete[] backE;
-        delete[] model;
-    }
-    else
-    {
-        // train both Transitions and Emissions
-        if (train)
-        {
-            trainParams(sig, kmerSeq, forM, forE, backM, backE, T, N, model, alphabetSize, numKmers, kmerSize);
-            std::cout << "Z:" << Zb << std::endl;
-            delete[] sig;
-            delete[] kmerSeq;
-            delete[] forM;
-            delete[] forE;
-            delete[] backM;
-            delete[] backE;
-            delete[] model;
-        }
-        else
-        {
-            delete[] sig;
-            delete[] kmerSeq;
-            delete[] model;
-            double *LPM = new double[TN];
-            logP(LPM, forM, backM, Zb, TN); // log probs for segmentation
-            delete[] forM;
-            delete[] backM;
-            double *LPE = new double[TN];
-            logP(LPE, forE, backE, Zb, TN); // log probs for extension
-            delete[] forE;
-            delete[] backE;
-            std::list<std::string> segString;
-            getBorders(segString, LPM, LPE, T, N, kmerSize);
-
-            for (const auto &seg : segString)
-            {
-                std::cout << seg;
-            }
-            std::cout << std::endl;
-
-            // calculate sum of segment probabilities
-            if (prob)
-            {
-                double sum;
-                for (std::size_t t = 0; t < T; ++t)
-                {
-                    sum = -INFINITY;
-                    for (std::size_t n = 0; n < N; ++n)
-                    {
-                        sum = logPlus(sum, LPM[t * N + n]);
-                    }
-                    std::cout << sum << ",";
-                }
-                std::cout << std::endl;
-
-                // to plot probability heatmap
-                // for (std::size_t t = 0; t < T; ++t)
-                // {
-                //     for (std::size_t n = 0; n < N; ++n)
-                //     {
-                //         std::cout << exp(logPlus(LPM[t * N + n], LPE[t * N + n])) << ",";
-                //     }
-                //     std::cout << "\n";
-
-                //     for (std::size_t n = 0; n < N; ++n)
-                //     {
-                //         std::cout << logPlus(forM[t * N + n], forE[t * N + n]) << ",";
-                //     }
-                //     std::cout << "\n";
-
-                //     for (std::size_t n = 0; n < N; ++n)
-                //     {
-                //         std::cout << logPlus(backM[t * N + n], backE[t * N + n]) << ",";
-                //     }
-                //     std::cout << "\n";
-                // }
-                // std::cout << std::flush;
-            }
-            // Clean up
-            delete[] LPM;
-            delete[] LPE;
-        }
-    }
-}
-#endif
