+++ src/cpp/NTC.cpp
@@ -3,28 +3,11 @@
 // github: https://github.com/JannesSP
 // website: https://jannessp.github.io
 
-#include "dynamont_NTC.hpp"
+#include "NTC.hpp"
 
 std::size_t TNK, NK;
 double ppTNm, ppTNe, ppTKm, ppTKe;
 int alphabetSize, kmerSize, halfKmerSize, stepSize;
-bool rna;
-std::unordered_map<std::string, double> transitions_NTK = {
-    {"a1", -1.0},
-    {"a2", -1.0},
-    {"p1", -1.0},
-    {"p2", -1.0},
-    {"p3", -1.0},
-    {"s1", -1.0},
-    {"s2", -1.0},
-    {"s3", -1.0},
-    {"e1", -1.0},
-    {"e2", -1.0},
-    {"e3", -1.0},
-    {"e4", -1.0},
-    {"i1", -1.0},
-    {"i2", -1.0},
-    {"e1", -1.0}};
 
 /**
  * Computes the posterior probabilities for a hidden Markov model (HMM) using
@@ -922,281 +905,6 @@ void traceback(std::size_t t, std::size_t n, std::size_t k,
 }
 
 /**
- * @brief Backtracing function for A state.
- *
- * This function backtraces the A state.
- * It checks which state was the previous state and calls the corresponding function.
- * If no match is found, it outputs an error message.
- *
- * @param t Current time index.
- * @param n Current segment index.
- * @param k Current kmer index.
- * @param APSEI The 2D array of scores.
- * @param logAPSEI The 2D array of log scores.
- * @param segString The list to store the segment strings.
- * @param K number of kemrs
- * @param segProb The vector to store the segment probabilities.
- *
- * @details
- * This function backtraces the A state.
- * It checks which state was the previous state and calls the corresponding function.
- * If no match is found, it outputs an error message.
- */
-// void funcA(const std::size_t t, const std::size_t n, const std::size_t k, std::unordered_map<std::size_t, std::array<dproxy, NUMMAT>> &APSEI, std::unordered_map<std::size_t, std::array<dproxy, NUMMAT>> &logAPSEI, std::list<std::string> &segString, const std::size_t K, std::vector<double> &segProb)
-// {
-//     const std::size_t currentIdx = t * NK + n * K + k;
-//     const std::size_t prevIdx = (t - 1) * NK + (n - 1) * K;
-//     // Cache the sc value to avoid redundant lookups
-//     const double sc = APSEI[currentIdx][0];
-//     const double logScore = logAPSEI[currentIdx][0];
-//     segProb.push_back(exp(logScore));
-//     if (t == 1 && n == 1)
-//     {
-//         // Start value
-//         segString.push_front("M" + std::to_string(halfKmerSize) + "," + std::to_string(0) + "," + formattedMedian(segProb) + "," + itoa(k, alphabetSize, kmerSize, rna) + ";");
-//         return;
-//     }
-//     for (std::size_t preKmer = precessingKmer(k, 0, stepSize, alphabetSize); preKmer < K; preKmer += stepSize)
-//     {
-//         if (t > 1 && n > 1)
-//         {
-//             // Check match with E state
-//             if (sc == APSEI[prevIdx + preKmer][3] + logScore)
-//             {
-//                 segString.push_front("M" + std::to_string(n - 1 + halfKmerSize) + "," + std::to_string(t - 1) + "," + formattedMedian(segProb) + "," + itoa(k, alphabetSize, kmerSize, rna) + ";");
-//                 segProb.clear();
-//                 return funcE(t - 1, n - 1, preKmer, APSEI, logAPSEI, segString, K, segProb);
-//             }
-//             // Check match with I state
-//             if (sc == APSEI[prevIdx + preKmer][4] + logScore)
-//             {
-//                 segString.push_front("M" + std::to_string(n - 1 + halfKmerSize) + "," + std::to_string(t - 1) + "," + formattedMedian(segProb) + "," + itoa(k, alphabetSize, kmerSize, rna) + ";");
-//                 segProb.clear();
-//                 return funcI(t - 1, n - 1, preKmer, APSEI, logAPSEI, segString, K, segProb);
-//             }
-//         }
-//     }
-//     // If no match is found, output an error
-//     std::cerr << "Error in backtracing funcA!: t: " << t << ", n: " << n << ", k: " << k << "\n";
-// }
-
-/**
- * Backtracing function for E state.
- *
- * @param t Current time index.
- * @param n Current segment index.
- * @param k Current kmer index.
- * @param APSEI The 2D array of scores.
- * @param logAPSEI The 2D array of log scores.
- * @param segString The list to store the segment strings.
- * @param K number of k-mers
- * @param segProb The vector to store the segment probabilities.
- *
- * @details
- * This function backtraces the E state.
- * It checks which state was the previous state and calls the corresponding function.
- * If no match is found, it outputs an error message.
- */
-// void funcE(const std::size_t t, const std::size_t n, const std::size_t k, std::unordered_map<std::size_t, std::array<dproxy, NUMMAT>> &APSEI, std::unordered_map<std::size_t, std::array<dproxy, NUMMAT>> &logAPSEI, std::list<std::string> &segString, const std::size_t K, std::vector<double> &segProb)
-// {
-//     const std::size_t currentIdx = NK * t + n * K + k;
-//     const std::size_t prevIdx = NK * (t - 1) + n * K + k;
-//     // Cache the sc value to avoid redundant lookups
-//     const double sc = APSEI[currentIdx][3];
-//     const double logScore = logAPSEI[currentIdx][3];
-//     segProb.push_back(exp(logScore));
-//     if (t > 1 && n > 0)
-//     {
-//         // Check match with A state
-//         if (sc == APSEI[prevIdx][0] + logScore)
-//         {
-//             return funcA(t - 1, n, k, APSEI, logAPSEI, segString, K, segProb);
-//         }
-//         // Check match with E state
-//         if (sc == APSEI[prevIdx][3] + logScore)
-//         {
-//             return funcE(t - 1, n, k, APSEI, logAPSEI, segString, K, segProb);
-//         }
-//         // Check match with S state
-//         if (sc == APSEI[prevIdx][2] + logScore)
-//         {
-//             return funcS(t - 1, n, k, APSEI, logAPSEI, segString, K, segProb);
-//         }
-//         // Check match with P state
-//         if (sc == APSEI[prevIdx][1] + logScore)
-//         {
-//             return funcP(t - 1, n, k, APSEI, logAPSEI, segString, K, segProb);
-//         }
-//     }
-//     else [[unlikely]]
-//     { // Start value with t==0 && n==0
-//         segString.push_front("M" + std::to_string(halfKmerSize) + ",0," + formattedMedian(segProb) + "," + itoa(k, alphabetSize, kmerSize, rna) + ";");
-//         segProb.clear();
-//         return;
-//     }
-//     // If no match is found, output an error
-//     std::cerr << "Error in backtracing funcE!: t: " << t << ", n: " << n << ", k: " << k << "\n";
-// }
-
-/**
- * @brief Backtraces and constructs segmentation for the P state.
- *
- * This function performs backtracing to construct the segmentation string
- * for the P (probationary) state, using scores from the forward iteration
- * stored in APSEI and logAPSEI. It also updates the probabilities for each
- * segment.
- *
- * @param t Current time step.
- * @param n Current sequence index.
- * @param k Current kmer index.
- * @param APSEI Map containing forward scores.
- * @param logAPSEI Map containing logarithmic forward scores.
- * @param segString List to store the segmentation strings.
- * @param K Number of kmers.
- * @param segProb Vector to store probabilities for each segment.
- */
-// void funcP(const std::size_t t, const std::size_t n, const std::size_t k, std::unordered_map<std::size_t, std::array<dproxy, NUMMAT>> &APSEI, std::unordered_map<std::size_t, std::array<dproxy, NUMMAT>> &logAPSEI, std::list<std::string> &segString, const std::size_t K, std::vector<double> &segProb)
-// {
-//     // Precompute commonly used indices and values
-//     const std::size_t currentIdx = t * NK + n * K + k;
-//     const double sc = APSEI[currentIdx][1];
-//     const double logScore = logAPSEI[currentIdx][1];
-//     const std::size_t prevBaseIdx = NK * (t - 1) + n * K; // Common base index for previous time step
-//     segProb.push_back(exp(logScore));
-//     if (t == 1 && n > 0)
-//     {
-//         // Start value
-//         segString.push_front("P" + std::to_string(halfKmerSize) + "," + std::to_string(0) + "," + formattedMedian(segProb) + "," + itoa(k, alphabetSize, kmerSize, rna) + ";");
-//         return;
-//     }
-//     if (t > 1 && n > 0)
-//     {
-//         for (std::size_t preKmer = precessingKmer(k, 0, stepSize, alphabetSize); preKmer < K; preKmer += stepSize)
-//         {
-//             const std::size_t prevIdx = prevBaseIdx + preKmer;
-//             // Check match with E state
-//             if (sc == APSEI[prevIdx][3] + logScore)
-//             {
-//                 segString.push_front("P" + std::to_string(n - 1 + halfKmerSize) + "," + std::to_string(t - 1) + "," + formattedMedian(segProb) + "," + itoa(k, alphabetSize, kmerSize, rna) + ";");
-//                 segProb.clear();
-//                 return funcE(t - 1, n, preKmer, APSEI, logAPSEI, segString, K, segProb);
-//             }
-//             // Check match with S state
-//             if (sc == APSEI[prevIdx][2] + logScore)
-//             {
-//                 segString.push_front("P" + std::to_string(n - 1 + halfKmerSize) + "," + std::to_string(t - 1) + "," + formattedMedian(segProb) + "," + itoa(k, alphabetSize, kmerSize, rna) + ";");
-//                 segProb.clear();
-//                 return funcS(t - 1, n, preKmer, APSEI, logAPSEI, segString, K, segProb);
-//             }
-//             // Check match with I state
-//             if (sc == APSEI[prevIdx][4] + logScore)
-//             {
-//                 segString.push_front("P" + std::to_string(n - 1 + halfKmerSize) + "," + std::to_string(t - 1) + "," + formattedMedian(segProb) + "," + itoa(k, alphabetSize, kmerSize, rna) + ";");
-//                 segProb.clear();
-//                 return funcI(t - 1, n, preKmer, APSEI, logAPSEI, segString, K, segProb);
-//             }
-//         }
-//     }
-//     // If no match is found, output an error
-//     std::cerr << "Error in backtracing funcP!: t: " << t << ", n: " << n << ", k: " << k << "\n";
-// }
-
-/**
- * @brief Backtracing for S state
- *
- * @param[in] t time step
- * @param[in] n sequence index
- * @param[in] k kmer index
- * @param[in] APSEI map of scores from forward iteration
- * @param[in] logAPSEI map of log scores from forward iteration
- * @param[out] segString list of strings representing the segmentation
- * @param[in] K number of kmers
- * @param[out] segProb vector of probabilities for each segment
- *
- * This function is used to construct the segmentation string by backtracing
- * the most likely state path from the forward iteration scores.
- */
-// void funcS(const std::size_t t, const std::size_t n, const std::size_t k, std::unordered_map<std::size_t, std::array<dproxy, NUMMAT>> &APSEI, std::unordered_map<std::size_t, std::array<dproxy, NUMMAT>> &logAPSEI, std::list<std::string> &segString, const std::size_t K, std::vector<double> &segProb)
-// {
-//     const std::size_t currentIdx = NK * t + n * K + k;
-//     const std::size_t prevIdx = NK * (t - 1) + (n - 1) * K + k;
-//     // Cache sc and logScore to avoid repeated map lookups
-//     const double sc = APSEI[currentIdx][2];
-//     const double logScore = logAPSEI[currentIdx][2];
-//     segProb.push_back(exp(logScore));
-//     if (t == 1 && n == 1)
-//     {
-//         // Start value
-//         // segString.push_front("S" + std::to_string(halfKmerSize) + "," + std::to_string(0) + "," + itoa(k, alphabet_size, kmerSize, rna) + ";");
-//         return;
-//     }
-//     if (t > 1 && n > 1)
-//     {
-//         // Check match with E state
-//         if (sc == APSEI[prevIdx][3] + logScore)
-//         {
-//             // segString.push_front("S" + std::to_string(n-1+halfKmerSize) + "," + std::to_string(t-1) + "," + itoa(k, alphabet_size, kmerSize, rna) + ";");
-//             return funcE(t - 1, n - 1, k, APSEI, logAPSEI, segString, K, segProb);
-//         }
-//         // Check match with P state
-//         if (sc == APSEI[prevIdx][1] + logScore)
-//         {
-//             // segString.push_front("S" + std::to_string(n-1+halfKmerSize) + "," + std::to_string(t-1) + "," + itoa(k, alphabet_size, kmerSize, rna) + ";");
-//             return funcP(t - 1, n - 1, k, APSEI, logAPSEI, segString, K, segProb);
-//         }
-//         // Check match with I state
-//         if (sc == APSEI[prevIdx][4] + logScore)
-//         {
-//             // segString.push_front("S" + std::to_string(n-1+halfKmerSize) + "," + std::to_string(t-1) + "," + itoa(k, alphabet_size, kmerSize, rna) + ";");
-//             return funcI(t - 1, n - 1, k, APSEI, logAPSEI, segString, K, segProb);
-//         }
-//     }
-//     // If no match is found, output an error
-//     std::cerr << "Error in backtracing funcS!: t: " << t << ", n: " << n << ", k: " << k << "\n";
-// }
-
-/**
- * @brief Backtracing function for the Insertion (I) state
- *
- * @param t The current time step
- * @param n The current position
- * @param k The current kmer
- * @param APSEI The forward probabilities table
- * @param logAPSEI The forward log probabilities table
- * @param segString The output segment string
- * @param K number of kmers
- * @param segProb The output segment probabilities
- *
- * @return void
- */
-// void funcI(const std::size_t t, const std::size_t n, const std::size_t k, std::unordered_map<std::size_t, std::array<dproxy, NUMMAT>> &APSEI, std::unordered_map<std::size_t, std::array<dproxy, NUMMAT>> &logAPSEI, std::list<std::string> &segString, const std::size_t K, std::vector<double> &segProb)
-// {
-//     const std::size_t currentIdx = NK * t + n * K + k;
-//     const std::size_t prevIdx = NK * t + (n - 1) * K + k;
-//     // Cache the score and logScore to avoid repeated lookups
-//     const double sc = APSEI[currentIdx][4];
-//     const double logScore = logAPSEI[currentIdx][4];
-//     segProb.push_back(exp(logScore));
-//     if (t > 0 && n > 1)
-//     {
-//         // Check match with I state
-//         if (sc == APSEI[prevIdx][4] + logScore)
-//         {
-//             // segString.push_front("I" + std::to_string(n-1+halfKmerSize) + "," + std::to_string(t-1) + "," + std::to_string(exp(logScore)) + "," + itoa(k, alphabet_size, kmerSize, rna) + ";");
-//             return funcI(t, n - 1, k, APSEI, logAPSEI, segString, K, segProb);
-//         }
-//         // Check match with E state
-//         if (sc == APSEI[prevIdx][3] + logScore)
-//         {
-//             // segString.push_front("I" + std::to_string(n-1+halfKmerSize) + "," + std::to_string(t-1) + "," + std::to_string(exp(logScore)) + "," + itoa(k, alphabet_size, kmerSize, rna) + ";");
-//             return funcE(t, n - 1, k, APSEI, logAPSEI, segString, K, segProb);
-//         }
-//     }
-//     // If no match is found, output an error
-//     std::cerr << "Error in backtracing funcI!: t: " << t << ", n: " << n << ", k: " << k << "\n";
-// }
-
-/**
  * Computes transition probabilities for a hidden Markov model (HMM) using
  * forward and backward probabilities, signal data, and k-mer sequences.
  *
@@ -1467,250 +1175,3 @@ void trainParams(const double *sig, const int *kmerSeq, std::unordered_map<std::
     delete[] newMeans;
     delete[] newStdevs;
 }
-
-/**
- * Main entry point for the dynamont 3d sparsed command line tool.
- *
- * @param argc Number of arguments passed to the program.
- * @param argv Array of strings containing the arguments passed to the program.
- *
- * @return 0 if the program succeeded, 3 if the Z values did not match between forward and backward matrices, 4 if the signal input is missing, 5 if the read input is missing.
- *
- * @details
- * This function parses command line arguments, sets up the transition parameters and model based on the specified pore type, reads the input signal and nucleotide sequence, and processes them to calculate segmentation probabilities. Depending on the arguments, it either calculates the partition function Z, trains the transition and emission parameters, or outputs the alignment with segment border probabilities.
- */
-#ifndef UNIT_TESTING
-int main(int argc, char *argv[])
-{
-    // speedup for I/O
-    std::ios_base::sync_with_stdio(0);
-    std::cin.tie(0);
-    std::cout.tie(0);
-
-    bool train, calcZ, prob; // atrain
-    std::string pore, modelpath;
-
-    // std::cerr precisions
-    std::cerr << std::fixed << std::showpoint << std::setprecision(5);
-    // std::cerr precisions
-    std::cout << std::fixed << std::showpoint << std::setprecision(5);
-
-    argparse::ArgumentParser program("dynamont 3d sparsed", "0.1");
-    program.add_argument("-a1", "--alignscore1").help("Transition parameter").default_value(-1.0).scan<'g', double>().store_into(transitions_NTK["a1"]);     // a1
-    program.add_argument("-a2", "--alignscore2").help("Transition parameter").default_value(-1.0).scan<'g', double>().store_into(transitions_NTK["a2"]);     // a2
-    program.add_argument("-e1", "--extendscore1").help("Transition parameter").default_value(-1.0).scan<'g', double>().store_into(transitions_NTK["e1"]);    // e1
-    program.add_argument("-e2", "--extendscore2").help("Transition parameter").default_value(-1.0).scan<'g', double>().store_into(transitions_NTK["e2"]);    // e2
-    program.add_argument("-e3", "--extendscore3").help("Transition parameter").default_value(-1.0).scan<'g', double>().store_into(transitions_NTK["e3"]);    // e3
-    program.add_argument("-e4", "--extendscore4").help("Transition parameter").default_value(-1.0).scan<'g', double>().store_into(transitions_NTK["e4"]);    // e4
-    program.add_argument("-s1", "--sequencescore1").help("Transition parameter").default_value(-1.0).scan<'g', double>().store_into(transitions_NTK["s1"]);  // s1
-    program.add_argument("-s2", "--sequencescore2").help("Transition parameter").default_value(-1.0).scan<'g', double>().store_into(transitions_NTK["s2"]);  // s2
-    program.add_argument("-s3", "--sequencescore3").help("Transition parameter").default_value(-1.0).scan<'g', double>().store_into(transitions_NTK["s3"]);  // s2
-    program.add_argument("-p1", "--polishscore1").help("Transition parameter").default_value(-1.0).scan<'g', double>().store_into(transitions_NTK["p1"]);    // p1
-    program.add_argument("-p2", "--polishscore2").help("Transition parameter").default_value(-1.0).scan<'g', double>().store_into(transitions_NTK["p2"]);    // p2
-    program.add_argument("-p3", "--polishscore3").help("Transition parameter").default_value(-1.0).scan<'g', double>().store_into(transitions_NTK["p3"]);    // p2
-    program.add_argument("-i1", "--insertionscore1").help("Transition parameter").default_value(-1.0).scan<'g', double>().store_into(transitions_NTK["i1"]); // i1
-    program.add_argument("-i2", "--insertionscore2").help("Transition parameter").default_value(-1.0).scan<'g', double>().store_into(transitions_NTK["i2"]); // i2
-    program.add_argument("--train").help("Switch algorithm to transition and emission parameter training mode").default_value(false).implicit_value(true).store_into(train);
-    program.add_argument("-z", "--calcZ").help("Switch algorithm to only calculate Z").default_value(false).implicit_value(true).store_into(calcZ);
-    program.add_argument("-m", "--model").help("Path to kmer model table").default_value("/home/yi98suv/projects/dynamont/data/norm_models/rna_r9.4_180mv_70bps.model").store_into(modelpath);
-    program.add_argument("-r", "--pore").help("Pore used to sequence the data").choices("rna_r9", "dna_r9", "rna_rp4", "dna_r10_260bps", "dna_r10_400bps").store_into(pore);
-    // unused, just here to match the other modes
-    program.add_argument("-p", "--probabilty").help("Print out the segment border probability").default_value(false).implicit_value(true).store_into(prob); //.store_into(prob);
-    program.add_argument("-t").help("Number of threads to use").default_value(1).scan<'i', int>();
-
-    program.parse_args(argc, argv);
-
-    omp_set_dynamic(1);
-    omp_set_num_threads(program.get<int>("-t"));
-
-    // load default and set parameters
-    if (pore == "rna_r9")
-    {
-        kmerSize = 5;
-        rna = true;
-        // taken from the trained NT version of dynamont
-        ppTNm = log(NT_rna_r9_transitions.at("m1")), ppTNe = log(NT_rna_r9_transitions.at("e2"));
-        ppTKm = log(NT_rna_r9_transitions.at("m1")), ppTKe = log(NT_rna_r9_transitions.at("e2"));
-        updateTransitions(NTK_rna_r9_transitions, transitions_NTK);
-    }
-    else if (pore == "dna_r9")
-    {
-        kmerSize = 5;
-        rna = false;
-        // taken from the trained NT version of dynamont
-        ppTNm = log(NT_dna_r9_transitions.at("m1")), ppTNe = log(NT_dna_r9_transitions.at("e2"));
-        ppTKm = log(NT_dna_r9_transitions.at("m1")), ppTKe = log(NT_dna_r9_transitions.at("e2"));
-        updateTransitions(NTK_dna_r9_transitions, transitions_NTK);
-    }
-    else if (pore == "rna_rp4")
-    {
-        kmerSize = 9;
-        rna = true;
-        // taken from the trained NT version of dynamont
-        ppTNm = log(NT_rna_rp4_transitions.at("m1")), ppTNe = log(NT_rna_rp4_transitions.at("e2"));
-        ppTKm = log(NT_rna_rp4_transitions.at("m1")), ppTKe = log(NT_rna_rp4_transitions.at("e2"));
-        updateTransitions(NTK_rna_rp4_transitions, transitions_NTK);
-    }
-    else if (pore == "dna_r10_260bps")
-    {
-        kmerSize = 9;
-        rna = false;
-        ppTNm = log(NT_dna_r10_260bps_transitions.at("m1")), ppTNe = log(NT_dna_r10_260bps_transitions.at("e2"));
-        ppTKm = log(NT_dna_r10_260bps_transitions.at("m1")), ppTKe = log(NT_dna_r10_260bps_transitions.at("e2"));
-        updateTransitions(NTK_dna_r10_260bps_transitions, transitions_NTK);
-    }
-    else if (pore == "dna_r10_400bps")
-    {
-        kmerSize = 9;
-        rna = false;
-        ppTNm = log(NT_dna_r10_400bps_transitions.at("m1")), ppTNe = log(NT_dna_r10_400bps_transitions.at("e2"));
-        ppTKm = log(NT_dna_r10_400bps_transitions.at("m1")), ppTKe = log(NT_dna_r10_400bps_transitions.at("e2"));
-        updateTransitions(NTK_dna_r10_400bps_transitions, transitions_NTK);
-    }
-    halfKmerSize = kmerSize / 2;
-
-    checkModelpath(modelpath);
-    auto result = readKmerModel(modelpath, kmerSize, rna);
-    std::tuple<double, double> *model = std::get<0>(result);
-    alphabetSize = std::get<1>(result);
-    // polishing dimension K = number of possible kmers
-    const std::size_t K = std::get<2>(result);
-
-    stepSize = pow(alphabetSize, kmerSize - 1);
-    std::string signal, read;
-
-    // echo 107,107,107.2,108.0,108.9,111.2,105.7,104.3,107.1,105.7,105,105 CAAAAA| src\segment.exe
-    // read input, signal and read whitespace separated in single line
-    getline(std::cin, signal);
-    getline(std::cin, read);
-
-    // exit if wrong input ...
-    if (signal.empty())
-    {
-        std::cout << "Signal missing!" << std::endl;
-        exit(4);
-    }
-    else if (read.empty())
-    {
-        std::cout << "Read missing!" << std::endl;
-        exit(5);
-    }
-
-    // process signal T: convert std::string to double std::array
-    const std::size_t T = count(signal.begin(), signal.end(), ',') + 2; // len(sig) + 1
-    checkInput(T, read.size(), kmerSize);
-    const std::size_t N = read.size() - kmerSize + 1 + 1; // N is number of kmers in sequence + 1
-    // std::cerr << "T: " << T << ", " << "N: " << N << ", " << "K: " << K << ", " << "inputsize: " << TNK << "\n";
-    NK = N * K;
-    TNK = T * NK;
-
-    double *sig = new double[T - 1];
-    std::string value;
-    std::stringstream ss(signal);
-    int i = 0;
-    while (getline(ss, value, ','))
-    {
-        sig[i++] = stod(value);
-    }
-    // process read N: convert std::string to int std::array
-
-    int *kmerSeq = new int[N - 1];
-    for (std::size_t n = 0; n < N - 1; ++n)
-    {
-        kmerSeq[n] = kmer2int(read.substr(n, kmerSize), alphabetSize);
-    }
-
-    // deallocate memory
-    ss.clear();
-    signal.erase();
-    read.erase();
-    value.erase();
-
-    std::vector<std::size_t> allowedKeys = preProcTNK(sig, kmerSeq, T, N, K, model);
-    // std::cerr<<"dense: "<<allowedKeys.size()/double(TNK)<<" ("<<allowedKeys.size()<<" / "<<TNK-allowedKeys.size()<<")"<<"\n"; //", sparse: "<<1-(allowedKeys.size()/double(TNK))<<" ("<<TNK-allowedKeys.size()<<")"<<"\n";
-    std::unordered_map<std::size_t, std::array<dproxy, NUMMAT>> forAPSEI;
-
-    // std::cerr<<"forward"<<"\n";
-    logF(sig, kmerSeq, forAPSEI, allowedKeys, K, model);
-    // std::cerr<<"backward"<<"\n";
-    std::unordered_map<std::size_t, std::array<dproxy, NUMMAT>> backAPSEI;
-    logB(sig, kmerSeq, backAPSEI, allowedKeys, T, N, K, model);
-
-    double Zf = -INFINITY;
-    double Zb = -INFINITY;
-    for (std::size_t k = 0; k < K; ++k)
-    {
-        Zf = logPlus(Zf, forAPSEI[TNK - 1 - k][3]);
-        Zb = logPlus(Zb, backAPSEI[k][3]);
-    }
-
-    // Numeric error is scaled by input size, Z in forward and backward should match by some numeric error EPSILON
-    if (abs(Zf - Zb) / TNK >= EPSILON || std::isinf(Zf) || std::isinf(Zb))
-    {
-        std::cerr << "Z values between matrices do not match! forZ: " << Zf << ", backZ: " << Zb << ", " << abs(Zf - Zb) / TNK << " > " << EPSILON << std::endl;
-        delete[] sig;
-        delete[] kmerSeq;
-        delete[] model;
-        exit(3);
-    }
-
-    // std::cerr<<"Zf: "<<Zf<<", Zb: "<<Zb<<", "<<abs(Zf-Zb)/TNK<<" <! "<<EPSILON<<"\n";
-
-    if (calcZ)
-    {
-        std::cout << Zf << std::endl;
-    }
-    else
-    {
-        std::unordered_map<std::size_t, std::array<dproxy, NUMMAT>> logAPSEI;
-        logP(logAPSEI, forAPSEI, backAPSEI, Zf, allowedKeys);
-
-        // train both Transitions and Emissions
-        if (train)
-        {
-            trainParams(sig, kmerSeq, forAPSEI, backAPSEI, logAPSEI, allowedKeys, T, N, K, model);
-            std::cout << "Z:" << Zf << std::endl;
-
-            // print out segmentation std::string
-        }
-        else
-        {
-            // Alignment output
-            std::list<std::string> segString;
-            getBorders(segString, logAPSEI, allowedKeys, T, N, K);
-
-            for (const auto &seg : segString)
-            {
-                std::cout << seg;
-            }
-            std::cout << std::endl;
-
-            // calculate sum of segment border probabilities
-            if (prob)
-            {
-                double sum = -INFINITY;
-                std::size_t lastT = T, t;
-                for (const std::size_t &tnk : allowedKeys)
-                {
-                    t = tnk / NK;
-                    if (t != lastT)
-                    {
-                        lastT = t;
-                        std::cout << sum << ",";
-                        sum = -INFINITY;
-                    }
-                    for (const int &i : {0, 1})
-                    { // sum up prob for new segment in dimension K
-                        sum = logPlus(sum, logAPSEI.at(tnk)[i]);
-                    }
-                }
-                std::cout << std::endl;
-            }
-        }
-    }
-    delete[] sig;
-    delete[] kmerSeq;
-    delete[] model;
-    return 0;
-}
-#endif
